refactor(settings): left-sidebar IA + unified save + destructive confirm (#153 PR-B)
Complete settings page redesign — replaces the 3-tab layout with a left-
sidebar 7-category navigation, unified sticky save bar, destructive-change
confirmation dialogs, and cross-panel navigation guard.

## IA restructure: 3 tabs → 7 sidebar categories

Old: General / Features / Advanced (top tabs, each with own save logic)
New: General / Storage / Camera Access / Streaming / AI Detection /
     Recording & Processing / Advanced (left sidebar, unified save)

| Category | Content |
|----------|---------|
| General | Timezone + interface preferences |
| Storage | Retention days + disk threshold + WebDAV |
| Camera Access | Auto-discover (toggle + credentials) |
| Streaming | RTMP port + SRT port |
| AI Detection | Global AI + confidence + MiBeeVision keys + per-camera AI + ROI |
| Recording & Processing | Merge toggle + transcoding (FFmpeg) |
| Advanced | Resource estimates (informational) |

Mobile: sidebar collapses to horizontal scrollable icon tabs.

## Unified save: 4 mechanisms → 1

- Bottom sticky save bar appears when ANY panel is dirty
- settingsForm coordinator aggregates dirty state across all panels
- Single Save button calls saveAll() which saves each dirty panel
- Reset button restores all panels to last-saved state
- Navigation guard (hashchange interceptor) covers all panels

## Destructive-change confirmation

Before saving, the coordinator collects destructive warnings from all panels:
- Reducing retention → 'recordings older than N days will be deleted'
- Disabling WebDAV → 'file managers lose access'
- Changing WebDAV RW mode → 'camera upload affected'
- Disabling RTMP/SRT → 'push streams rejected'
- Disabling auto-discover → 'cameras no longer auto-added'
- Disabling AI → 'real-time detection stops'
- Disabling merge → 'timeline becomes fragmented'

All warnings shown in a single confirmation dialog.

## Files

New: Settings.svelte (shell with sidebar + sticky save + guards)
New: settings-form.svelte.ts (extended with getDestructiveWarnings + reactivity)
New: 7 Panel components (GeneralPanel, StoragePanel, CameraAccessPanel,
     StreamingPanel, AIPanel, ProcessingPanel, AdvancedPanel)
Deleted: GeneralTab.svelte, FeaturesTab.svelte, AdvancedTab.svelte
Updated: en.json + zh.json (sidebar labels + destructive warning messages)

Each Panel registers with settingsForm on mount, owns its form state +
API load/save, and exposes isDirty/save/reset/getDestructiveWarning.
