+++ web/src/lib/i18n/en.json
@@ -1092,6 +1092,22 @@
   "settings.timezoneLocal": "Server timezone",
   "settings.timezoneSaved": "Timezone updated",
   "settings.title": "Settings",
+  "settings.reset": "Reset",
+  "settings.sidebar.general": "General",
+  "settings.sidebar.storage": "Storage",
+  "settings.sidebar.cameras": "Camera Access",
+  "settings.sidebar.streaming": "Streaming",
+  "settings.sidebar.ai": "AI Detection",
+  "settings.sidebar.processing": "Recording & Processing",
+  "settings.sidebar.advanced": "Advanced",
+  "settings.destructive.retentionReduce": "Retention period reduced — recordings older than {{days}} days will be deleted immediately.",
+  "settings.destructive.webdavDisable": "WebDAV will be disabled — external file managers will lose access to recordings.",
+  "settings.destructive.webdavRWChange": "WebDAV read-write mode changed — camera upload behavior will be affected.",
+  "settings.destructive.rtmpOff": "RTMP ingest will be disabled — external push streams will be rejected.",
+  "settings.destructive.srtOff": "SRT listener will be disabled — external SRT streams will be rejected.",
+  "settings.destructive.autodiscoverOff": "Auto-discovery will be disabled — new cameras will no longer be added automatically.",
+  "settings.destructive.aiOff": "AI detection will be disabled — real-time detection on all cameras will stop.",
+  "settings.destructive.mergeOff": "Segment merging will be disabled — timeline will become fragmented with short clips.",
   "settings.unsavedChanges": "Unsaved changes",
   "settings.unsavedDiscard": "Discard & Leave",
   "settings.unsavedMessage": "You have unsaved changes that will be lost. Discard changes and leave?",
