+++ web/src/lib/i18n/zh.json
@@ -1092,6 +1092,22 @@
   "settings.timezoneLocal": "服务器时区",
   "settings.timezoneSaved": "时区已更新",
   "settings.title": "设置",
+  "settings.reset": "重置",
+  "settings.sidebar.general": "通用",
+  "settings.sidebar.storage": "存储",
+  "settings.sidebar.cameras": "摄像头接入",
+  "settings.sidebar.streaming": "直播",
+  "settings.sidebar.ai": "AI 检测",
+  "settings.sidebar.processing": "录像与处理",
+  "settings.sidebar.advanced": "高级",
+  "settings.destructive.retentionReduce": "保留期限缩短 — 超过 {{days}} 天的录像将被立即删除。",
+  "settings.destructive.webdavDisable": "WebDAV 将被关闭 — 外部文件管理器将无法访问录像。",
+  "settings.destructive.webdavRWChange": "WebDAV 读写模式已更改 — 摄像头上传行为将受到影响。",
+  "settings.destructive.rtmpOff": "RTMP 接入将被关闭 — 外部推流将被拒绝。",
+  "settings.destructive.srtOff": "SRT 接收器将被关闭 — 外部 SRT 流将被拒绝。",
+  "settings.destructive.autodiscoverOff": "自动发现将被关闭 — 新摄像头加入网络将不再自动添加。",
+  "settings.destructive.aiOff": "AI 检测将被关闭 — 所有摄像头的实时检测将停止。",
+  "settings.destructive.mergeOff": "段合并将被关闭 — 时间轴将变为碎片化的短片段。",
   "settings.unsavedChanges": "未保存的修改",
   "settings.unsavedDiscard": "放弃并离开",
   "settings.unsavedMessage": "您有未保存的修改，离开后将丢失。确认放弃修改并离开？",
